+++ autosubmit/platforms/paramiko_platform.py
@@ -839,7 +839,6 @@ class ParamikoPlatform(Platform):
                 f'check_job() The job id ({job_id}) is not an integer neither a string.')
             job.new_status = job_status
         sleep_time = 5
-        sleep(2)
         self.send_command(self.get_check_job_cmd(job_id))
         while self.get_ssh_output().strip(" ") == "" and retries > 0:
             retries = retries - 1
@@ -957,7 +956,6 @@ class ParamikoPlatform(Platform):
         job_list_cmd = self.parse_job_list(job_list)
         cmd = self.get_check_all_jobs_cmd(job_list_cmd)
         sleep_time = 5
-        sleep(sleep_time)
         slurm_error = False
         e_msg = ""
         try:
@@ -1215,7 +1213,7 @@ class ParamikoPlatform(Platform):
                 stdout = chan.makefile('rb', bufsize)
                 stderr = chan.makefile_stderr('rb', bufsize)
                 return stdin, stdout, stderr
-            except (paramiko.SSHException, ConnectionError, socket.error) as e:
+            except (paramiko.SSHException, ConnectionError, socket.error, IOError) as e:
                 Log.warning(f'A networking error occurred while executing command [{command}]: {str(e)}')
                 if not self.connected or not self.transport or not self.transport.active:
                     self.restore_connection(None)